kloud/start: add retry mechanism to make it more stable
